Folder structure has been improved
