+++ Tomi.Application/Features/ShoppingCarts/Commands/AddItem/AddItemCommand.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using Tomi.Application.Models;
+
+namespace Tomi.Application.Features.ShoppingCarts.Commands.AddItem
+{
+    public class AddItemCommand : IRequest<ShoppingCartItemModel>
+    {
+        public string UserId { get; set; }
+        public string ProductId { get; set; }
+    }
+}
