+++ Tomi.Application/ServiceRegistration.cs
@@ -3,10 +3,14 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
 using Tomi.Application.Auth.Commands;
+using Tomi.Application.Features.Coupons.Commands.Apply;
+using Tomi.Application.Features.Coupons.Queries.GetAll;
+using Tomi.Application.Features.Products.Queries.GetAll;
+using Tomi.Application.Features.Products.Queries.GetById;
+using Tomi.Application.Features.ShoppingCarts.Commands.AddItem;
+using Tomi.Application.Features.ShoppingCarts.Commands.RemoveAllItem;
+using Tomi.Application.Features.ShoppingCarts.Queries.GetAll;
 using Tomi.Application.Mappings;
-using Tomi.Application.Services.Handlers.Coupons;
-using Tomi.Application.Services.Handlers.Products;
-using Tomi.Application.Services.Handlers.ShoppingCarts;
 
 namespace Tomi.Application
 {
@@ -22,9 +26,9 @@ namespace Tomi.Application
             services.AddMediatR(typeof(AuthenticateCommandHandler).GetTypeInfo().Assembly);
             services.AddMediatR(typeof(RegisterCommandHandler).GetTypeInfo().Assembly);
 
-			services.AddMediatR(typeof(AddItemToShoppingCart).GetTypeInfo().Assembly);
-            services.AddMediatR(typeof(RemoveAllItemsFromShoppingCart).GetTypeInfo().Assembly);
-			services.AddMediatR(typeof(GetAllShoppingCartItems).GetTypeInfo().Assembly);
+			services.AddMediatR(typeof(AddItemToShoppingCartHandler).GetTypeInfo().Assembly);
+            services.AddMediatR(typeof(ClearShoppingCartHandler).GetTypeInfo().Assembly);
+			services.AddMediatR(typeof(GetAllShoppingCartItemsByUserIdHandler).GetTypeInfo().Assembly);
 
 			services.AddMediatR(typeof(ApplyCouponHandler).GetTypeInfo().Assembly);
 			services.AddMediatR(typeof(GetAllCouponsHandler).GetTypeInfo().Assembly);
