+++ Tomi.Application/Services/TokenService.cs
@@ -1,13 +1,16 @@
 ﻿using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using System;
+using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
 using System.Security.Claims;
 using System.Text;
+using System.Threading.Tasks;
 using Tomi.Domain.Entities;
-using Tomi.Infrastructure.Settings;
+using Tomi.Domain.Settings;
 
-namespace Tomi.Infrastructure.Services
+namespace Tomi.Application.Services
 {
 	public class TokenService
 	{
