+++ Tomi.Domain/Settings/JwtSettings.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Tomi.Infrastructure.Settings
+namespace Tomi.Domain.Settings
 {
 	public class JwtSettings
 	{
