+++ Tomi.WebAPI/Controllers/ShoppingCartController.cs
@@ -2,7 +2,11 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Tomi.Application.Features.ShoppingCarts.Commands;
+using Tomi.Application.Features.ShoppingCarts.Commands.AddItem;
+using Tomi.Application.Features.ShoppingCarts.Commands.RemoveAllItem;
+using Tomi.Application.Features.ShoppingCarts.Commands.RemoveItem;
 using Tomi.Application.Features.ShoppingCarts.Queries;
+using Tomi.Application.Features.ShoppingCarts.Queries.GetAll;
 using Tomi.Domain.Entities;
 
 namespace Tomi.WebAPI.Controllers
@@ -39,19 +43,12 @@ namespace Tomi.WebAPI.Controllers
 		}
 
         [HttpDelete("clear-shopping-cart")]
-        public async Task<IActionResult> RemoveAllItemsFromShoppingCart()
+        public async Task<IActionResult> ClearShoppingCart()
         {
-			var command = new RemoveAllItemsCommand();
+			var command = new ClearShoppingCartCommand();
 			command.UserId = UserId;
 			return Ok(await Mediator.Send(command));
         }
-        [HttpDelete("remove-item-from-shopping-cart-compeletely")]
-        public async Task<IActionResult> RemoveItemFromShoppingCartCompeletely(string productId)
-        {
-            var command = new RemoveItemCompeletelyCommand();
-            command.UserId = UserId;
-            command.ProductId = productId;
-            return Ok(await Mediator.Send(command));
-        }
+
     }
 	}
